bots/TeleAssistantBot/requirements.txt
python-telegram-bot[rate-limiter]==20.7
PyYAML==6.0
python-dotenv==0.21.0
pydub==0.25.1
charade==1.0.3
pymongo>=4.4.0
openai==1.6.0
tiktoken>=0.3.0
pydub==0.25.1 # Manipulate audio with an simple and easy high level interface.